Remove api_utils.py
